Casting to the correct type for X11 calls.

+++ src/map_tif.c
@@ -1180,7 +1180,16 @@ void draw_geotiff_image_map (Widget w,
         GTIFFree (gtif);
         XTIFFClose (tif);
         // Update to screen
-        (void)XCopyArea(XtDisplay(da),pixmap,XtWindow(da),gc,0,0,screen_width,screen_height,0,0);
+        (void)XCopyArea(XtDisplay(da),
+            pixmap,
+            XtWindow(da),
+            gc,
+            0,
+            0,
+            (unsigned int)screen_width,
+            (unsigned int)screen_height,
+            0,
+            0);
         return;
     }
 
@@ -1345,7 +1354,16 @@ Samples Per Pixel: 1
             GTIFFree (gtif);
             XTIFFClose (tif);
             // Update to screen
-            (void)XCopyArea(XtDisplay(da),pixmap,XtWindow(da),gc,0,0,screen_width,screen_height,0,0);
+            (void)XCopyArea(XtDisplay(da),
+                pixmap,
+                XtWindow(da),
+                gc,
+                0,
+                0,
+                (unsigned int)screen_width,
+                (unsigned int)screen_height,
+                0,
+                0);
             return;
         }
